Move arduino stuff to separate folder
