+++ EvilAppleJuice-ESP32-INO/devices.hpp
